xml-backend: Don't unlink lock file if NFS link check fails
If the NFS link check results in too many links to the lock file,
this implies that another process elsewhere has acquired the lock.

Do not delete the lock file in this scenario. Depending on the
exact timing involved this may result in neither process acquiring
the lock, leaving a stale lock file.