Bug 798695 - Deleting everything from the "Transfer" cell after suggestions pop-up restricts search to the first 30 accounts
The combo box list is capable of displaying all accounts if expanded
without starting searching, so changing the search string to "" shouldn't
only return a small subset of accounts with no way to fix it.

Skip the search and return all accounts in the combo box as it normally
does if the account list is opened without searching for something.