Bug 798564 - GnuCash is slow when there are a lot of open tabs/registers
Every time a change to a transaction is saved by navigating away from it,
the UI freezes for almost a second. This is because
gnc_split_register_move_cursor() calls gnc_gui_refresh_internal() which
refreshes all the registers and I have 35+ registers open.

Register refreshes should be deferred on tabs that are not visible until
those tabs are focused again.

Use focus information to suppress the last part of the refresh_handler()
if the ledger isn't currently visible. When it becomes visible again, run
the deferred refresh step.