Bug 798564 - GnuCash is slow when there are a lot of open tabs/registers
When loading the book on startup, GnuCash takes a long time. This is partly
because gnc_restore_all_state() is loading data for every register page and
I have 35+ registers open.

Register loads should be deferred on tabs that are not visible until those
tabs are focused, like what already happens for restored open reports.

Don't automatically focus pages when they are recreated. This has to
disregard a preference to get them to be in the right order so it is
specifically only for recreates (and not a "background" or "without focus"
way of creating pages).

Change the register page so that it doesn't assume it will be in focus on
creation. Don't immediately load or refresh the ledger (this will happen
the first time it is focused instead).

The previously active page is still focused automatically when loading is
finished and this will load the ledger if it's a register.

There is still a separate unresolved issue (bug 798942) that the accounts
tree is taking 2 seconds to load 1500 accounts if the page is visible. The
page is only visible when opening (or re-opening) the book while GnuCash is
running. This commit improves the behaviour when the accounts tree is not
the first page in a window because it will not be focused while it's being
loaded.