Bug 777472 - reconcile does not work if transaction selected
If there's an existing transaction with pending edits when starting
reconciliation, that transaction can appear in the list of transactions and
be modified by the reconciliation but as soon as navigation to another
transaction occurs the pending edit will be saved, overwriting the change
to mark it as reconciled. The result is that reconciliation appears to have
failed or not marked the transaction as reconciled.

This commonly happens when marking several transactions as cleared before
performing reconciliation. If the last transaction is still being edited
it will typically be saved after the reconciliation finishes and overwrite
any change made.

Potential modification or removal of transactions affecting reconciliation
could be detected by looking for pending edits to transactions involving
the account being reconciled but it's not trivial to determine if a pending
edit in another register would add the transaction to the account being
reconciled.

Check that there's no outstanding activity in any register pages before
starting a reconciliation. It is still possible to start modifying a
transaction after the reconciliation window is open but this will stop the
most common issue with the process.