Register preference "Jump action for multiple splits"
Make the behaviour when the "Jump" operation encounters a transaction with
more than two splits configurable.

This can be configured as:
* Do nothing (this is the default)
* Use the split with the largest value
* Use the split with the smallest value

This makes it much easier to jump between accounts for stock transactions
that include splits for fees and taxes, or dividend transactions that have
a zero value split for the stock account.