Disallow using "Jump" on an expanded split for the current account
If you've selected a split for this account, for consistency with selecting
the split of another account we should do nothing. You're already on the
account for the split you selected.

Previously it would still go to the "other" account but that's going to
make the "multiple split" options confusing.