Bug 777472 - reconcile does not work if transaction selected
If there's an existing transaction with pending edits when starting
reconciliation, that transaction can appear in the list of transactions and
be modified by the reconciliation but as soon as navigation to another
transaction occurs the pending edit will be saved, overwriting the change
to mark it as reconciled. The result is that reconciliation appears to have
failed or not marked the transaction as reconciled.

This commonly happens when marking several transactions as cleared before
performing reconciliation. If the last transaction is still being edited
it will typically be saved after the reconciliation finishes and overwrite
any change made.

Check that there's no outstanding activity in the current register page
before starting a reconciliation.

It is still possible to start modifying a transaction after the
reconciliation window is open but this will stop the most common issue with
the process.

Starting a reconciliation from the account tree is left unprotected.