Bug 798564 - GnuCash is slow to start when there are a lot of open registers
When loading the book on startup, GnuCash takes a long time. This is partly
because gnc_restore_all_state() is loading data for every register page and
I have 35+ registers open.

Register loads should be deferred on tabs that are not visible until those
tabs are focused, like what already happens for restored open reports.

Don't immediately load or refresh the ledger (this will happen the first
time it is focused instead).