Fix uninstall target for cmake > 3.25
execute_process and exce_program were changed in cmake 3.26 to run
only at configure time, breaking cmake_uninstall.cmake. The recommened
replacement is add_custom_target and add_custom_commsnd, but there's
no way to run a loop in either of those so this commit replaces
cmake_uninstall.cmake with cmake_uninstall.sh. This works on Microsoft
Windows beacuse at present we have to build on Windows in a MSYS2
shell. If at some point we gain the ability to build in a straight
Windows environment then we'll need a Windows CMD equivalent.